Remove unnecessary type casts
Commit UUID: 678ee3dd-cefd-4b4e-95c2-41637334e426

+++ trunk/blkdev.c
@@ -239,7 +239,7 @@ get_blkdev_size(int fd, uint64_t *sz)
     count *= size;
 #endif
 
-    *sz = (uint64_t)count;
+    *sz = count;
     return 0;
 }
